Prevent starting replication directly on the leader index alias.
(This was previously resulting in NPE)

The rationale is to avoid complexity for now since an alias can be
associated with multiple indices, although an extension could be to
allow replication if there is a single match.

Added an integration test to validate the case.

While at it, also expanded BasicReplicationIT to incorporate force merge
and ensure that doesn't impact replication.

Signed-off-by: Gopala Krishna Ambareesh <gopalak@amazon.com>