use proper gradle dep. to fetch security plugin
there's no need to manually construct a URL and try to fetch the
artefact from there. the plugin is published on maven and we can
directly use the gradle/maven dependency to fetch it.

in order to more easily identify it, a new configuration type has been
added: `opensearchPlugin`.

for reference see the discussions on the following issues/PRs:
- opensearch-project/cross-cluster-replication#274
- opensearch-project/cross-cluster-replication#455

Signed-off-by: Ralph Ursprung <Ralph.Ursprung@avaloq.com>